% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OurSlope.R
\name{OurSlope}
\alias{OurSlope}
\title{It performs the slope heuristic for the selection of a penalty constant}
\usage{
OurSlope(contrast, grille, penalty)
}
\arguments{
\item{contrast}{the Frobenius norm of X-M_est for all the value of the grid \code{grille}}

\item{grille}{the ordered grid of potential values for the penalty constant}

\item{penalty}{the penalty calculated for each value of the grid \code{grille}}
}
\value{
\code{Model_Selected} the selected parameter
}
\description{
It performs the slope heuristic for the selection of a penalty constant
}
